Moving firewall samples to a new folder to match agreed structure for upcoming sample work.
