+++ compute/cloud-client/firewall/composer.json
@@ -1,0 +1,6 @@
+{
+    "require": {
+        "google/cloud-compute": "^1.0.0",
+        "google/cloud-storage": "^1.26"
+    }
+}
