+++ compute/cloud-client/firewall/src/delete_firewall_rule.php
